# backend/docs/swagger.yaml
definitions:
...
  models.Banner:
    properties:
      created_at:
        description: 创建时间
        type: string
      effective_active:
        type: boolean
      end_time:
        description: 活动结束时间（可为空）
        type: string
      id:
        description: 主键（自增ID）
        type: integer
      image_url:
        description: 图片URL
        type: string
      in_time_window:
        type: boolean
      link_url:
        description: 跳转链接（可为空）
        type: string
      manual_active:
        type: boolean
      sort:
        description: 排序，数字越小越靠前
        type: integer
      start_time:
        description: 活动开始时间（可为空）
        type: string
      status:
        description: 状态：active, inactive
        type: string
      subtitle:
        description: 副标题（可为空）
        type: string
      title:
        description: 标题
        type: string
      updated_at:
        description: 更新时间
        type: string
    type: object
...
  utils.ErrorInfo:
    properties:
      code:
        description: 错误代码
        type: string
      message:
        description: 错误消息
        type: string
    type: object
...
  utils.Response:
    properties:
      data:
        description: 返回的数据
      error:
        allOf:
        - $ref: '#/definitions/utils.ErrorInfo'
        description: 错误信息
      message:
        description: 提示消息
        type: string
      success:
        description: 是否成功
        type: boolean
    type: object
host: 127.0.0.1:18080
info:
  contact:
    email: support@gohotel.com
    name: API Support
  description: 这是一个酒店预订管理系统的 RESTful API 服务
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 酒店管理系统 API
  version: "1.0"
paths:
  /api/admin/banners:
    get:
      consumes:
      - application/json
      description: 获取所有活动横幅，支持分页
      parameters:
      - description: 页码，默认1
        in: query
        name: page
        type: integer
      - description: 每页条数，默认10
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 获取所有活动横幅
      tags:
      - 活动管理
    post:
      consumes:
      - multipart/form-data
      description: 创建新的活动横幅，使用临时图片URL
      parameters:
      - description: 活动横幅标题
        in: formData
        name: title
        required: true
        type: string
      - description: 活动横幅副标题
        in: formData
        name: subtitle
        type: string
      - description: 临时图片URL（通过通用上传接口获取）
        in: formData
        name: temp_url
        required: true
        type: string
      - description: 点击跳转链接
        in: formData
        name: link_url
        type: string
      - description: 展示顺序
        in: formData
        name: sort
        type: integer
      - description: 活动开始时间
        in: formData
        name: start_time
        type: string
      - description: 活动结束时间
        in: formData
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Banner'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 创建活动横幅
      tags:
      - 活动管理
  /api/admin/banners/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取活动横幅详情
      parameters:
      - description: 活动横幅ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Banner'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 根据ID获取活动横幅
      tags:
      - 活动管理
    post:
      consumes:
      - multipart/form-data
      description: 更新活动横幅信息，使用临时图片URL
      parameters:
      - description: 活动横幅ID
        in: path
        name: id
        required: true
        type: string
      - description: 活动横幅标题
        in: formData
        name: title
        type: string
      - description: 活动横幅副标题
        in: formData
        name: subtitle
        type: string
      - description: 临时图片URL（通过通用上传接口获取）
        in: formData
        name: temp_url
        type: string
      - description: 点击跳转链接
        in: formData
        name: link_url
        type: string
      - description: 展示顺序
        in: formData
        name: sort
        type: integer
      - description: 活动开始时间
        in: formData
        name: start_time
        type: string
      - description: 活动结束时间
        in: formData
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Banner'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 更新活动横幅
      tags:
      - 活动管理
  /api/admin/banners/{id}/delete:
    post:
      consumes:
      - application/json
      description: 删除指定ID的活动横幅
      parameters:
      - description: 活动横幅ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 删除活动横幅
      tags:
      - 活动管理
...
  /api/banners/active:
    get:
      consumes:
      - application/json
      description: 获取激活状态的活动横幅，用于前端展示
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Banner'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取激活的活动横幅
      tags:
      - 活动管理
...
securityDefinitions:
  Bearer:
    description: 输入 "Bearer {token}" 格式的 JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"